We're updating our Privacy Policy
ClassDojo's mission is to connect teachers, students, and families to help them improve education in their communities. Protecting your privacy is fundamental to that mission.

We updated our Privacy Policy to reflect changes we made to strengthen your rights, increase transparency for how we protect your information, and to comply with the European Union’s General Data Protection Regulation (GDPR). These changes are effective May 23, 2018.

We encourage you to read our policies in full, but here are some highlights of what’s changed:

- We include a “Privacy Highlights” section at the top of our Privacy Policy to help you quickly understand how we collect, use, share, and protect your personal information
- We created a simple chart, linked from our Privacy Policy, to help you understand the information we have about you, how we use it, and where it is stored
- We created a short, easy-to-understand video to help students understand our Privacy Policy
- We highlighted your privacy rights, including how to exercise them, such as how to delete your personal information

Your continued use of ClassDojo after today is an acknowledgment that you agree to the updated policies.